@@ -1,17 +1,17 @@
 name: flutter_vegvesen
 description: Flutter/Dart package til å hente kjøretøyopplysninger fra Statens Vegvesen
-version: 1.0.2
+version: 1.1.0
 homepage: https://github.com/beerstorm-net/statens-vegvesen
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
-  flutter: ">=1.17.0 <2.0.0"
+  sdk: '>=2.12.0 <3.0.0'
+  flutter: ">=1.17.0"
 
 dependencies:
   flutter:
     sdk: flutter
 
-  http: ^0.12.2
+  http: ^0.13.0
 
 dev_dependencies:
   flutter_test:
